chore: demo de refacto de composant
