+++ app/src/utils/index.js
@@ -116,4 +116,8 @@ export const getDistance = (lat1, lon1, lat2, lon2) => {
   }
 };
 
+export function capitalizeFirstLetter(string) {
+  if (string) return string.charAt(0).toUpperCase() + string.slice(1);
+}
+
 export const regexPhoneFrenchCountries = `(\\+(33|590|594|262|596|269|687|689|508|681)|06|07)(?:\\W*\\d){8}$`;
